refactor: refactored load data mem to project pattern; add data mem debugger, some fixes and more. final sprint 1 🎉:

+++ src/types.h
@@ -42,7 +42,7 @@ typedef struct {
 
 typedef struct {
     int8_t data[256];
-    uint8_t size;
+    int size;
 } MemData;
 
 // Estado (útil para a função de back)
